Adapt the code to use GUI
