+++ src/main/java/duke/gui/MainWindow.java
@@ -11,7 +11,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 
 /**
- * Controller for Launcher.duke.gui.MainWindow. Provides the layout for the other controls.
+ * Controller for duke.Launcher.duke.gui.MainWindow. Provides the layout for the other controls.
  */
 public class MainWindow extends AnchorPane {
     @FXML
